Refactor keyboard navigation and input focus handling in completion components
- Moved keyboard arrow key support for scrolling from Input component to useCompletion hook for better organization.
- Added keyboard shortcut (Cmd+K / Ctrl+K) for toggling 'keepEngaged' mode with input focus after toggle.
- Enhanced input focus handling after AI responses and screenshot completions for improved user experience.
- Updated Input component UI to include a tooltip for the keyboard shortcut.
